pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secsignal"
version = "0.1.0"
description = "Production-grade Agentic RAG system for SEC financial intelligence"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    # Snowflake
    "snowflake-connector-python>=3.6.0",
    "snowflake-snowpark-python>=1.21.0",

    # LangGraph agent framework
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # Data processing
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",

    # GCS
    "google-cloud-storage>=2.18.0",

    # Observability
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "pre-commit>=4.0.0",
]
dbt = [
    "dbt-snowflake>=1.8.0",
]
frontend = [
    # Next.js frontend managed via npm in secsignal/frontend/
    # No Python dependencies required — kept for documentation
]
airflow = [
    "apache-airflow>=2.10.0",
    "apache-airflow-providers-snowflake>=5.7.0",
    "apache-airflow-providers-google>=10.23.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["secsignal"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
